requirements.txt
docling>=2.48.0
camelot-py>=1.0.9
opencv-python-headless>=4.10.0.84
pytesseract>=0.3.13
Pillow>=10.4.0
pdf2image>=1.17.0
sentence-transformers>=3.0.1
transformers>=4.44.2
torch>=2.8.0
faiss-cpu>=1.12.0
gradio>=4.44.1
numpy>=2.1.1
pandas>=2.2.2
rich>=13.8.1
